+++ Mapwize/Mapwize/Classes/MWZVenue.m
@@ -1,4 +1,5 @@
 #import "MWZVenue.h"
+#import "MWZGeometryFactory.h"
 
 @implementation MWZVenue
 
@@ -7,7 +8,15 @@
     _identifier = [dic objectForKey:@"_id"];
     _name = [dic objectForKey:@"name"];
     _alias = [dic objectForKey:@"alias"];
+    _defaultLanguage = [dic objectForKey:@"defaultLanguage"];
+    _supportedLanguages = [dic objectForKey:@"supportedLanguages"];
+    _icon = [dic objectForKey:@"icon"];
+    _geometry = [MWZGeometryFactory geometryWithDictionary:[dic objectForKey:@"geometry"]];
     _data = [dic objectForKey:@"data"];
+    _marker = [[MWZCoordinate alloc] initWithDictionary:[dic objectForKey:@"marker"]];
+    _isPublished = [[dic objectForKey:@"isPublished"] boolValue];
+    _areQrcodesDeployed = [[dic objectForKey:@"areQrcodesDeployed"] boolValue];
+    _areIbeaconsDeployed = [[dic objectForKey:@"areIbeaconsDeployed"] boolValue];
     return self;
 }